requirements.txt
mysql-connector-python
pyjwt
redis
